+++ src/index.js
@@ -22,6 +22,16 @@ export { default as LayoutContainer } from "./components/LayoutContainer/LayoutC
 export { ActionButton } from "./components/ActionButton/ActionButton";
 
 /**
+ * buttons
+ */
+export { Button } from "./components/Button/Button";
+
+/**
  * Heading with text link
  */
 export { Heading } from "./components/Heading/Heading";
+
+/**
+ * Image
+ */
+export { Image } from "./components/Image/Image";