+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Locked eslint version to 7.32.0 due to an error with jest
 - Heading component
 - Documentation and styling for Dividers, Colors, Typography
+- Button test cases/stories/html (react.js)
 - Image Component
 - Link Component